Balancing change; Enabled creation of random Atoms based on generation;

+++ Assets/Scripts/Features/Planets/Components/PlanetAtomModel.cs
@@ -8,6 +8,6 @@ public class PlanetAtomModel
     public string Name;
     public int AtomicNumber;
     public int Stock;
-    public float HarvestRate;
+    public int HarvestRate;
     public int UpgradeLevel;
 }
